+++ djangocms_navigation/cms_config.py
@@ -93,7 +93,11 @@ class NavigationCMSAppConfig(CMSAppConfig):
     )
     navigation_models = {
         # model_class : field(s) to search in menu item form UI
-        Page: ["pagecontent_set__title__icontains"]
+        Page: [
+            "pagecontent_set__title__icontains",
+            "urls__slug__icontains",
+            "urls__path__icontains",
+        ]
     }
     djangocms_references_enabled = True
     reference_fields = [