+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 Unreleased
 ==========
+* fix: Navigation nodes with a page attached display edit urls in edit mode
 * feat: When adding a MenuItem the Page objects are filtered by PageUrl slug and path
 
 1.5.1 (2022-08-26)